// apps/blog/config/types.ts
export interface PostMatter {
  title: string
  date: Date
  dateString: string
  thumbnail: string
  desc: string
}

export interface Post extends PostMatter {
  url: string
  slug: string
  categoryPath: string
  content: string
  readingMinutes: number
  categoryPublicName: string
}

export interface CategoryDetail {
  dirName: string
  publicName: string
  count: number
}

export interface HeadingItem {
  text: string
  link: string
  indent: number
}